requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
gunicorn==20.1.0
idna==2.10
psycopg2==2.9.1
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.2.0